Backend/app/requirements.txt
Django==3.2.9
psycopg2-binary==2.9.1
architect
djangorestframework
django-cors-headers
